refactor: merge the renderToMountedElement into the render
