+++ src/runtime/render.ts
@@ -1,7 +1,7 @@
-import { Canvas as GCanvas } from '@antv/g';
+import { Canvas as GCanvas, DisplayObject, Group } from '@antv/g';
 import { Renderer as CanvasRenderer } from '@antv/g-canvas';
 import { Plugin as DragAndDropPlugin } from '@antv/g-plugin-dragndrop';
-import { debounce, deepMix } from '@antv/util';
+import { debounce, deepMix, isEmpty } from '@antv/util';
 import { createLibrary } from '../stdlib';
 import { select } from '../utils/selection';
 import { emitEvent, CHART_LIFE_CIRCLE, offEvent } from '../utils/event';
@@ -90,6 +90,35 @@ export function render<T extends G2ViewTree = G2ViewTree>(
 
   // Return the container HTML element wraps the canvas or svg element.
   return normalizeContainer(canvas.getConfig().container);
+}
+
+export function renderToMountedElement<T extends G2ViewTree = G2ViewTree>(
+  options: T,
+  context: G2Context = {},
+  callback?: () => void,
+): DisplayObject {
+  // Initialize the context if it is not provided.
+  const { width = 640, height = 480, on } = options;
+  const keyed = inferKeys(options);
+  const { library = createLibrary(), group = new Group() } = context;
+
+  if (isEmpty(group?.parentElement)) {
+    throw new Error(`Unmounted group`);
+  }
+
+  const selection = select(group);
+  context.group = group;
+  context.library = library;
+
+  emitEvent(on, CHART_LIFE_CIRCLE.BEFORE_RENDER);
+  // Plot the chart and mutate context.
+  // Make sure that plot chart after container is ready for every time.
+  plot<T>({ ...keyed, width, height }, selection, library)
+    .then(callback)
+    .then(() => emitEvent(on, CHART_LIFE_CIRCLE.AFTER_RENDER));
+
+  // Return the Group wraps the canvas or svg element.
+  return group;
 }
 
 export function destroy<T extends G2ViewTree = G2ViewTree>(
